+++ android/src/main/java/com/smileidentity/react/viewmanagers/SmileIDDocumentVerificationViewManager.kt
@@ -40,7 +40,6 @@ class SmileIDDocumentVerificationViewManager(
       view.documentType = it.getStringOrDefault("documentType", null)
       view.idAspectRatio = it.getFloatOrDefault("idAspectRatio")
       view.allowNewEnroll = it.getBoolOrDefault("allowNewEnroll", false)
-      view.skipApiSubmission = it.getBoolOrDefault("skipApiSubmission", false)
       view.useStrictMode = it.getBoolOrDefault("useStrictMode", false)
     }
   }