+++ android/src/main/java/com/smileidentity/react/viewmanagers/SmileIDSmartSelfieCaptureViewManager.kt
@@ -29,6 +29,7 @@ class SmileIDSmartSelfieCaptureViewManager(
       view.allowAgentMode = it.getBoolOrDefault("allowAgentMode", false)
       view.showAttribution = it.getBoolOrDefault("showAttribution", true)
       view.showInstructions = it.getBoolOrDefault("showInstructions", true)
+      view.useStrictMode = it.getBoolOrDefault("useStrictMode", false)
       view.showConfirmation =
         it.getBoolOrDefault("showConfirmation", true)
     }