+++ android/src/main/java/com/smileidentity/react/views/SmileIDSmartSelfieEnrollmentEnhancedView.kt
@@ -8,6 +8,7 @@ import com.smileidentity.SmileID
 import com.smileidentity.compose.SmartSelfieEnrollmentEnhanced
 import com.smileidentity.react.results.SmartSelfieCaptureResult
 import com.smileidentity.react.utils.SelfieCaptureResultAdapter
+import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.util.randomUserId
 import kotlinx.coroutines.CoroutineScope
@@ -15,8 +16,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 
-class SmileIDSmartSelfieEnrollmentEnhancedView(context: ReactApplicationContext) : SmileIDView(context) {
-  private val viewScope = CoroutineScope(Dispatchers.Main + Job())
+class SmileIDSmartSelfieEnrollmentEnhancedView(context: ReactApplicationContext) : SmileIDSelfieView(context) {
   override fun renderContent() {
     composeView.apply {
       setContent {
@@ -27,40 +27,10 @@ class SmileIDSmartSelfieEnrollmentEnhancedView(context: ReactApplicationContext)
             allowNewEnroll = allowNewEnroll ?: false,
             showAttribution = showAttribution,
             showInstructions = showInstructions,
+            skipApiSubmission = skipApiSubmission,
             extraPartnerParams = extraPartnerParams,
-          ) {res ->
-            viewScope.launch {
-              when (res) {
-                is SmileIDResult.Success -> {
-                  val result =
-                    SmartSelfieCaptureResult(
-                      selfieFile = res.data.selfieFile,
-                      livenessFiles = res.data.livenessFiles,
-                      apiResponse = res.data.apiResponse,
-                    )
-                  val newMoshi =
-                    SmileID.moshi
-                      .newBuilder()
-                      .add(SelfieCaptureResultAdapter.FACTORY)
-                      .build()
-                  val json =
-                    try {
-                      newMoshi
-                        .adapter(SmartSelfieCaptureResult::class.java)
-                        .toJson(result)
-                    } catch (e: Exception) {
-                      emitFailure(e)
-                      return@launch
-                    }
-                  json?.let { js ->
-                    emitSuccess(js)
-                  }
-                }
-
-                is SmileIDResult.Error -> emitFailure(res.throwable)
-              }
-            }
-          }
+            onResult = { res -> handleResultCallback(res) },
+          )
         }
       }
     }