+++ ios/View/SmileIDBiometricKYCView.swift
@@ -8,10 +8,9 @@ struct SmileIDBiometricKYCView: View, SmileIDFileUtilsProtocol {
     var smileIDUIViewDelegate: SmileIDUIViewDelegate
     var body: some View {
         NavigationView {
-            if let idInfo = product.idInfo, let consentInformation = product.consentInformation {
+            if let idInfo = product.idInfo {
                 SmileID.biometricKycScreen(
                     idInfo: idInfo,
-                    consentInformation: consentInformation, // already validated in the SmileIDBiometricKYCViewManager
                     userId: product.userId ?? generateUserId(),
                     jobId: product.jobId ?? generateJobId(),
                     allowNewEnroll: product.allowNewEnroll,
@@ -20,6 +19,7 @@ struct SmileIDBiometricKYCView: View, SmileIDFileUtilsProtocol {
                     showInstructions: product.showInstructions,
                     useStrictMode: product.useStrictMode,
                     extraPartnerParams: product.extraPartnerParams as [String: String],
+                    consentInformation: product.consentInformation, // already validated in the SmileIDBiometricKYCViewManager
                     delegate: self
                 )
             } else {