+++ ios/View/SmileIDSmartSelfieAuthView.swift
@@ -10,10 +10,12 @@ struct SmileIDSmartSelfieAuthView: View, SmileIDFileUtilsProtocol {
         NavigationView {
             SmileID.smartSelfieAuthenticationScreen(
                 userId: product.userId ?? generateUserId(),
+                jobId: product.jobId ?? generateJobId(),
                 allowNewEnroll: product.allowNewEnroll,
                 allowAgentMode: product.allowAgentMode,
                 showAttribution: product.showAttribution,
                 showInstructions: product.showInstructions,
+                skipApiSubmission: product.skipApiSubmission,
                 extraPartnerParams: product.extraPartnerParams as [String: String],
                 delegate: self
             )