+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Release Notes
+
+## 10.2.6
+* Added enhanced SmartSelfie™ capture Selfie capture screen component
+* Added `skipApiSubmission` to SmartSelfie™ capture which defaults to `false` and will allow Selfie capture without submission to the api
+* Make consent information optional on Biometric KYC, Enhanced KYC and Enhanced Document Verification
+* Bump android to 10.5.2 (https://github.com/smileidentity/android/releases/tag/v10.5.2)
+* Bump iOS to 10.4.2 (https://github.com/smileidentity/ios/releases/tag/v10.4.2)
+
 ## 10.2.5
 * Added enhanced SmartSelfie™ capture to docV, enhanced docV, and biometric kyc
 * Added consent information to BioMetric KYC and Enhanced Document Verification